Refractor platform repo + drop mingw/OSx + call arch consitently 64/32
before on linux it was x86 or x86_64 and win32/win64 on windows, now its always "32", or "64".

We drop: mingw support and osx support (both not been maintained).

Consequently we also drop any cross compilation efforts (but make it easier to integrate one afterwards, as long as someone recompiles the dependencies for mingw and maintains that stuff)

platform paths (CMAKE_lib/include/program_PATH) don't need to be added in a toolchain file -> normal include now.